# MM-9801 Performance improvements for large numbers of posts