--- a/src/store/initial_state.js
+++ b/src/store/initial_state.js
@@ -43,6 +43,7 @@ const state = {
         posts: {
             posts: {},
             postsInChannel: {},
+            postsInThread: {},
             selectedPostId: '',
             currentFocusedPostId: '',
             messagesHistory: {