… test/actions/posts.test.js
… test/selectors/posts.test.js