#### Summary
After some profiling with 5000 posts loaded locally, the major bottleneck was in the `formatPostInChannel` function.

<img width="962" alt="screen shot 2018-03-16 at 14 04 30" src="https://user-images.githubusercontent.com/2672098/37599893-3db974dc-2b5c-11e8-87d1-f0ba7190bd6d.png">

We were spending a ton of time in there looping through every single post we had locally to do two things:
* Check if the current user started or replied to the thread of the post in question (needed for notification purposes)
* Count the number of replies in the thread

While debugging, removing this one loop took switching to a channel with 5000 posts from taking ~60 seconds to just ~200ms.

To remove the loop without functionality loss, I added a new reducer `postsInThread`. It is very similar to the existing `postsInChannel` reducer, except instead of tracking posts in channels it was for tracking posts in threads.

Here's some profiling numbers showing the improvements on `formatPostInChannel` from doing that:

| Commit | Self Time | Total Time |
| --- | --- | --- |
| pre-changes | 30,735.6 ms | 94,823.6 ms |
| bc6c865 | 184.4 ms | 206.9 ms |

That reduced total time spent in that function by a factor of more than 450, which is quite the improvement. Even still, `formatPostInChannel` was still the performance bottleneck because it was being run on every single post in the channel, even if those posts weren't going to be used. I fixed this by adding another parameter to the `getPostsInChannel` selector called `numPosts` for specifying how many posts the selector should return. The results of that for time spent in `formatPostInChannel`:

 | Commit | Self Time | Total Time |
| --- | --- | --- |
| bc6c865 | 184.4 ms | 206.9 ms |
| f8ac2fe | 0.9 ms | 13.4 ms |

In addition to improving `formatPostInChannel`, I also went through and fixed every instance where we were looping through all the posts by using the `postsInThread` reducer.

This should be backwards compatible with old usages of the selectors (except for getPostsInChannel).

#### More Potential Improvements
This helped a lot but there is still more room for improvement:
* If a thread has 1000s of posts, we will still loop through them all for each post
* There is still a ton of loop invariant code contained in `formatPostInChannel` that could be pulled out to improve performance
* The selector recalculates very frequently, finding a way to reduce that would help a bunch
* We spend a lot of time sorting, both in the selectors and reducers for posts

#### Ticket Link
https://mattermost.atlassian.net/browse/MM-9801

#### Checklist
- [x] Added or updated unit tests (required for all new features)