@jwilander 2018-03-19
@enahum added it to initial state and I'll change the middleware when I submit the PR to update redux in the mobile repo