@koxen 2018-03-20
is anyone else seeing problems with "load new messages" not working after loading second page?
is it just coincidence that it started happening when I applied this patch?