@jwilander 2018-03-20
@koxen there needs to be webapp changes to support loading more changes, that's why it won't load past the second page for you

That will come in the web app PR to update mattermost-redux